Fixed Heading Issue with AHRS. Turns out mu-metal was working on preventing interference on compass but there was electrical noise on the accelerometer, applied a low pass filter on accelerometer readings.

+++ libraries/RR_CommonData/RR_CommonData.h
@@ -25,7 +25,7 @@ typedef struct RR_TelemetryOutgoingMessage_t
 	float Latitude;
 	float Longitude;
 	uint16_t DistanceToTarget;
-	int8_t Heading;
+	int16_t Heading;
 	int16_t	Bearing;
 	uint16_t Altitude;
